pyproject.toml
[project]
name = "smart-store-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.97.0",
    "openai-agents>=0.2.2",
    "python-dotenv>=1.1.1",
    "toolagent>=0.0.0",
]
